+++ BakeryPOS.API/BakeryPOS.API/DTOs/TopSellingProductDto.cs
@@ -4,7 +4,7 @@
     {
         public int ProductId { get; set; }
         public string ProductName { get; set; } = string.Empty;
-        public int TotalSold { get; set; }
+        public decimal TotalSold { get; set; }
         public decimal TotalRevenue { get; set; }
         public decimal GrowthPercentage { get; set; }
     }
